refactor(tui): extract the entry points and leave a facade behind
run_local, run_host, and run_guest move to tui::app, the legacy
single-pane runtime to tui::host, and raw-mode setup to tui::terminal.
src/tui.rs is now module declarations, re-exports, and a map of the
tree: 12,012 lines down to 65.
