+++ src/tui/terminal.rs
@@ -1,0 +1,64 @@
+//! Raw mode, the alternate screen, and the keyboard-enhancement handshake —
+//! entered on start and undone on drop.
+
+use std::{fs::OpenOptions, io, io::Write};
+
+use crossterm::{
+    event::{
+        DisableBracketedPaste, DisableMouseCapture, KeyboardEnhancementFlags,
+        PopKeyboardEnhancementFlags, PushKeyboardEnhancementFlags,
+    },
+    execute,
+    terminal::{LeaveAlternateScreen, disable_raw_mode},
+};
+
+pub(in crate::tui) struct TerminalGuard {
+    pub(in crate::tui) raw_mode: bool,
+    pub(in crate::tui) keyboard_enhancement: bool,
+    pub(in crate::tui) alternate_screen: bool,
+    pub(in crate::tui) bracketed_paste: bool,
+    pub(in crate::tui) mouse_capture: bool,
+}
+impl TerminalGuard {
+    pub(in crate::tui) fn new() -> Self {
+        Self {
+            raw_mode: false,
+            keyboard_enhancement: false,
+            alternate_screen: false,
+            bracketed_paste: false,
+            mouse_capture: false,
+        }
+    }
+}
+impl Drop for TerminalGuard {
+    fn drop(&mut self) {
+        let mut stdout = io::stdout();
+        if self.keyboard_enhancement {
+            let _ = execute!(stdout, PopKeyboardEnhancementFlags);
+            let _ = stdout.flush();
+            if let Ok(mut tty) = OpenOptions::new().write(true).open("/dev/tty") {
+                let _ = execute!(tty, PopKeyboardEnhancementFlags);
+                let _ = tty.flush();
+            }
+        }
+        if self.bracketed_paste {
+            let _ = execute!(stdout, DisableBracketedPaste);
+        }
+        if self.mouse_capture {
+            let _ = execute!(stdout, DisableMouseCapture);
+        }
+        if self.alternate_screen {
+            let _ = execute!(stdout, crossterm::cursor::Show, LeaveAlternateScreen);
+        }
+        if self.raw_mode {
+            let _ = disable_raw_mode();
+        }
+    }
+}
+pub(in crate::tui) fn enable_keyboard_enhancement() -> io::Result<bool> {
+    execute!(
+        io::stdout(),
+        PushKeyboardEnhancementFlags(KeyboardEnhancementFlags::DISAMBIGUATE_ESCAPE_CODES)
+    )?;
+    Ok(true)
+}
